LayoutTests/fast/images/image-orientation-none-local-expected.txt

PASS rotatedImage.offsetWidth is 50
PASS nonRotatedImage.offsetWidth is 100
The test should display two images, one rotated and one not.
